refactor(trip-details): Refactor TripDetails component and add new subcomponents
